move simready from toolkit to gensim and update relative docs
