+++ docs/source/features/generative_sim/index.rst
@@ -1,0 +1,9 @@
+Generative Simulation
+=====================
+
+Generative Simulation collects EmbodiChain features for generating simulation-ready assets and executing agent-driven task workflows.
+
+.. toctree::
+   :maxdepth: 2
+
+   SimReady Asset Pipeline <simready_pipeline.md>
