+++ docs/source/features/online_data.md
@@ -148,6 +148,6 @@ python examples/agents/datasets/online_dataset_demo.py
 
 ## See Also
 
-- [EmbodiAgent](agents.md) — Hierarchical agent that uses online data for training
+- [EmbodiAgent](generative_sim/agents.md) — Hierarchical agent that uses online data for training
 - [RL Architecture](../overview/rl/index.rst) — RL training pipeline
 - [Data Generation Tutorial](../tutorial/data_generation.rst) — Generating offline datasets
