+++ docs/source/index.rst
@@ -40,8 +40,7 @@ Table of Contents
    :glob:
 
    features/online_data.md
-   features/agents.md
-   features/simready_pipeline.md
+   features/generative_sim/index*
    features/workspace_analyzer/index*
    features/interaction/index*
    features/toolkits/index*
