+++ embodichain/gen_sim/simready_pipeline/core/asset.py
