+++ embodichain/gen_sim/simready_pipeline/parser/internal.py
@@ -19,8 +19,8 @@ import trimesh
 import pyrender
 from PIL import Image
 from pathlib import Path
-from embodichain.toolkits.simready_pipeline.core.asset import Asset
-from embodichain.toolkits.simready_pipeline.parser.base import AssetParser
+from embodichain.gen_sim.simready_pipeline.core.asset import Asset
+from embodichain.gen_sim.simready_pipeline.parser.base import AssetParser
 
 
 class InternalParser(AssetParser):
