+++ embodichain/gen_sim/simready_pipeline/parser/usd.py
@@ -19,9 +19,9 @@ from typing import Any, Dict
 
 import numpy as np
 import trimesh
-from embodichain.toolkits.simready_pipeline.parser.base import AssetParser
-from embodichain.toolkits.simready_pipeline.core.asset import Asset
-from embodichain.toolkits.simready_pipeline.utils.usd_utils import (
+from embodichain.gen_sim.simready_pipeline.parser.base import AssetParser
+from embodichain.gen_sim.simready_pipeline.core.asset import Asset
+from embodichain.gen_sim.simready_pipeline.utils.usd_utils import (
     convert_model_to_usd,
     DEFAULT_PHYSICS_PARAMS,
 )
