+++ embodichain/gen_sim/simready_pipeline/utils/texture_utils.py
